chore: refactor core settings loader and setter
