+++ Editor/UI/EditorWindows/SettingsWindow/SettingsEditor.cs
@@ -113,7 +113,7 @@ namespace ReadyPlayerMe.Core.Editor
         {
             AnalyticsEditorLogger.EventLogger.LogSetLoggingEnabled(evt.newValue);
             SDKLogger.EnableLogging(evt.newValue);
-            CoreSettingsUtil.SetEnableLogging(evt.newValue);
+            CoreSettingsSetter.SetEnableLogging(evt.newValue);
         }
 
         private void OnPrivacyPolicyClicked(MouseUpEvent evt)
