+++ Editor/Utils/CoreSettingsLoader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: da717eefac74935459a3f995092c5463
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
