+++ Editor/Utils/CoreSettingsSetter.cs
@@ -1,20 +1,10 @@
-using ReadyPlayerMe.Core.Data;
-using UnityEngine;
 using UnityEditor;
 
 namespace ReadyPlayerMe.Core.Editor
 {
-    [InitializeOnLoad]
-    public static class CoreSettingsUtil
+
+    public static class CoreSettingsSetter
     {
-        private const string PROJECT_RELATIVE_ASSET_PATH = "Assets/Ready Player Me/Resources/Settings/CoreSettings.asset";
-        private const string SETTINGS_SAVE_FOLDER = "Ready Player Me/Resources/Settings";
-
-        static CoreSettingsUtil()
-        {
-            EnsureSettingsExist();
-        }
-
         public static void SetEnableAnalytics(bool isEnabled)
         {
             CoreSettingsHandler.CoreSettings.EnableAnalytics = isEnabled;
@@ -44,23 +34,6 @@ namespace ReadyPlayerMe.Core.Editor
         {
             EditorUtility.SetDirty(CoreSettingsHandler.CoreSettings);
             AssetDatabase.SaveAssets();
-        }
-
-        public static void EnsureSettingsExist()
-        {
-            if (CoreSettingsHandler.CoreSettings == null)
-            {
-                CreateSettings();
-            }
-        }
-
-        private static void CreateSettings()
-        {
-            DirectoryUtility.ValidateDirectory($"{Application.dataPath}/{SETTINGS_SAVE_FOLDER}");
-            var newSettings = ScriptableObject.CreateInstance<CoreSettings>();
-            AssetDatabase.CreateAsset(newSettings, PROJECT_RELATIVE_ASSET_PATH);
-            AssetDatabase.SaveAssets();
-            AssetDatabase.Refresh();
         }
     }
 }
